# platform/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TerraGenius AI API",
    "description": "API for generating and managing infrastructure-as-code.\n\n## Authentication Flow\n\n1. User signs up or logs in using Firebase Authentication (email/password, Google, etc.) via the frontend.\n2. Frontend obtains a Firebase ID token.\n3. Frontend POSTs the ID token to `/auth/token`.\n4. Backend verifies the Firebase token and returns a custom JWT (`accessToken`).\n5. All protected API calls require `Authorization: Bearer <accessToken>`.\n6. The backend validates the JWT and attaches user context.\n\n**Public endpoints:** `/auth/token`, `/status`, `/health`, `/signup`, `/login`, `/reset-password`\n**Protected endpoints:** All `/api/` routes, `/generate`, `/deployments`, `/billing`, `/audit-logs`, etc.\n\n**Future Ready:** To swap auth providers (e.g. Auth0, Cognito), only the `/auth/token` logic and token validation need to change. The rest of the API and security model remains the same.\n\n**Testing:** Use the provided `test_auth_flow.js` script to verify the full login → token exchange → protected API flow.\n\n**Security:** Never send Firebase tokens directly to protected endpoints. Always use the custom JWT returned from `/auth/token`.\n",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Mock API Server"
    }
  ],
  "paths": {
...
    "/generations/{projectId}": {
      "get": {
        "summary": "List generations for a project",
        "description": "Retrieve all generation records for a specific project",
        "tags": ["generations"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "List of generations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create a new generation record",
        "description": "Create a new generation record in Firestore",
        "tags": ["generations"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Project identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Generation record created successfully"
          },
          "400": {
            "description": "Invalid request parameters"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/token endpoint"
      }
    },
    "schemas": {
      "GenerationRequest": {
        "type": "object",
        "properties": {
          "prompt": {
            "type": "string",
            "example": "Create an AWS S3 bucket for static website hosting."
          },
          "model": {
            "type": "string",
            "example": "gemini-2.5-flash-preview-04-17"
          },
          "provider": {
            "type": "string",
            "enum": ["aws", "gcp", "azure"],
            "example": "aws"
          }
        },
        "required": ["prompt", "model", "provider"]
      },
...
      "ValidationStepLog": {
        "type": "object",
        "properties": {
          "stage": {"type": "string", "example": "generate"},
          "status": {"type": "string", "example": "success"},
          "message": {"type": "string", "example": "AI response received."},
          "timestamp": {"type": "string", "format": "date-time"}
        },
        "required": ["stage", "status", "message", "timestamp"]
      },
...
      "GenerationListResponse": {
        "type": "object",
        "properties": {
          "generations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenerationResponse"
            }
          },
          "total": {
            "type": "integer"
          }
        },
        "required": ["generations", "total"]
      },
      "GenerationResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "prompt": { "type": "string" },
          "model": { "type": "string" },
          "provider": { "type": "string" },
          "status": { "type": "string" },
          "jobId": { "type": "string" },
          "files": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "language": { "type": "string" },
                "content": { "type": "string" }
              }
            }
          },
          "logs": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationStepLog" }
          },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        },
        "required": ["id", "userId", "projectId", "prompt", "model", "provider", "status", "jobId", "files", "logs", "createdAt", "updatedAt"]
      },
...
    }
  }
}